+++ rust/worker/src/execution/operator.rs
@@ -58,53 +58,3 @@ where
 {
     async fn run(&self, input: &I) -> O;
 }
-
-#[derive(Debug)]
-pub(super) struct PullLogsOperator {
-    client: Box<dyn Log>,
-}
-
-impl PullLogsOperator {
-    pub fn new(client: Box<dyn Log>) -> Self {
-        PullLogsOperator { client }
-    }
-}
-
-#[derive(Debug)]
-pub(super) struct PullLogsInput {
-    collection_id: Uuid,
-    offset: i64,
-    batch_size: i32,
-}
-
-impl PullLogsInput {
-    pub(super) fn new(collection_id: Uuid, offset: i64, batch_size: i32) -> Self {
-        PullLogsInput {
-            collection_id,
-            offset,
-            batch_size,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub(super) struct PullLogsOutput {
-    // TODO: Standardize on Vec<Box<EmbeddingRecord>> as our data chunk type and add a type alias
-    logs: Vec<Box<EmbeddingRecord>>,
-}
-
-#[async_trait]
-impl Operator<PullLogsInput, PullLogsOutput> for PullLogsOperator {
-    async fn run(&self, input: &PullLogsInput) -> PullLogsOutput {
-        let mut client_clone = self.client.clone();
-        let logs = client_clone
-            .read(
-                input.collection_id.to_string(),
-                input.offset,
-                input.batch_size,
-            )
-            .await
-            .unwrap();
-        PullLogsOutput { logs }
-    }
-}
